+++ load-balance/src/main/java/com/shieldblaze/expressgateway/loadbalance/l7/http/sessionpersistence/StickySessionSearchComparator.java
@@ -1,0 +1,38 @@
+/*
+ * This file is part of ShieldBlaze ExpressGateway. [www.shieldblaze.com]
+ * Copyright (c) 2020 ShieldBlaze
+ *
+ * ShieldBlaze ExpressGateway is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * ShieldBlaze ExpressGateway is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with ShieldBlaze ExpressGateway.  If not, see <https://www.gnu.org/licenses/>.
+ */
+package com.shieldblaze.expressgateway.loadbalance.l7.http.sessionpersistence;
+
+import com.shieldblaze.expressgateway.backend.Backend;
+
+import java.util.Comparator;
+
+public final class StickySessionComparator implements Comparator<Object> {
+
+    public static final StickySessionComparator INSTANCE = new StickySessionComparator();
+
+    private StickySessionComparator() {
+        // Prevent outside initialization
+    }
+
+    @Override
+    public int compare(Object o1, Object o2) {
+        String key = (String) o1;
+        Backend backend = (Backend) o2;
+        return backend.getHash().compareTo(key);
+    }
+}
