+++ load-balance/src/test/java/com/shieldblaze/expressgateway/loadbalance/l4/sessionpersistence/SourceIPHashTest.java
@@ -1,0 +1,4 @@
+import static org.junit.jupiter.api.Assertions.*;
+class SourceIPHashTest {
+  
+}
