+++ load-balance/src/test/java/com/shieldblaze/expressgateway/loadbalance/l7/http/sessionpersistence/StickySessionTest.java
@@ -1,0 +1,4 @@
+import static org.junit.jupiter.api.Assertions.*;
+class StickySessionTest {
+  
+}
